lisa.tests: load_tracking: Move *Migration out of staging
Move CPUMigration tests out of staging, into the main load_tracking.py
module since they have been stable for some time.
